Improve hydraulic heater step

+++ include/HydraulicHeater.h
@@ -15,14 +15,14 @@ class HydraulicHeater {
   void Update();
 
   // Power control (0-100%)
-  void SetPower(float power);
-  float GetPower() const { return power_; }
+  void SetPower(uint8_t power);
+  uint8_t GetPower() const { return power_; }
 
   // PWM output (0.0-1.0) with hardware mapping
   float GetOutput() const;
 
  private:
-  float power_;  // 0-100%
+  uint8_t power_;  // 0-100%
 
   // PWM hardware mapping constants
   static constexpr float kBandMin = 10.0f;  // Maximum speed PWM%
